requirements.txt
asgiref==3.2.10
astroid==2.4.0
colorama==0.4.3
Django==3.0
djangorestframework==3.12.1
emoji==0.5.4
ImageHash==4.1.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
numpy==1.19.4
Pillow==8.0.1
pylint==2.5.0
pytz==2020.1
PyWavelets==1.1.1
scipy==1.5.4
six==1.14.0
sqlparse==0.4.1
toml==0.10.0
wrapt==1.12.1
